refactor: rename folders to MVVM structure
