+++ app/src/main/java/com/example/android/view/ToiletActivity.java
@@ -1,4 +1,4 @@
-package com.example.android.activities;
+package com.example.android.view;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -9,26 +9,34 @@ import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.android.R;
-import com.example.android.adapters.RestroomListAdapter;
+import com.example.android.viewModel.ToiletListAdapter;
+import com.example.android.model.Restroom;
 import com.example.android.store.Store;
 
-public class MainActivity extends AppCompatActivity {
+public class ToiletActivity extends AppCompatActivity {
     private final Store store = Store.getInstance();
+    private Restroom restroom = null;
     private RecyclerView mRecyclerView;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        setContentView(R.layout.toilet_activity);
 
-        RestroomListAdapter mAdapter = new RestroomListAdapter(this, store.getRestrooms());
+        int restroomIndex = store.getShowingRestroomIndex();
+        if (restroomIndex != -1) {
+            restroom = store.getRestroom(restroomIndex);
+            setTitle(restroom.getLocation());
+        }
+
+        ToiletListAdapter mAdapter = new ToiletListAdapter(this, restroom.getToiletList());
         mRecyclerView = findViewById(R.id.recyclerView);
         mRecyclerView.setAdapter(mAdapter);
         mRecyclerView.setLayoutManager(new LinearLayoutManager(this));
     }
 
-    public void addNewRestroom(View view) {
-        Intent intent = new Intent(view.getContext(), AddRestroomActivity.class);
+    public void addNewToilet(View view) {
+        Intent intent = new Intent(view.getContext(), AddToiletActivity.class);
         startActivityForResult(intent, 1);
     }
