+++ app/src/main/java/com/example/android/viewModel/ToiletListAdapter.java
@@ -1,4 +1,4 @@
-package com.example.android.adapters;
+package com.example.android.viewModel;
 
 import android.content.Context;
 import android.graphics.Color;
@@ -16,8 +16,8 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.android.R;
-import com.example.android.classes.Toilet;
-import com.example.android.classes.ToiletState;
+import com.example.android.model.Toilet;
+import com.example.android.model.ToiletState;
 import com.example.android.store.Store;
 
 import java.util.ArrayList;
