+++ app/src/main/res/layout/add_restroom_activity.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.AddRestroomActivity">
+    tools:context=".view.AddRestroomActivity">
 
     <TextView
         android:id="@+id/restroomIdTextView"
