+++ app/src/main/res/layout/add_toilet_activity.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.AddToiletActivity">
+    tools:context=".view.AddToiletActivity">
 
     <TextView
         android:id="@+id/toiletIdTextView"
